% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perfect-foresight.R
\name{summary.dsge_perfect_foresight}
\alias{summary.dsge_perfect_foresight}
\title{Summary of Perfect Foresight Transition}
\usage{
\method{summary}{dsge_perfect_foresight}(object, ...)
}
\arguments{
\item{object}{A \code{dsge_perfect_foresight} object.}

\item{...}{Additional arguments (unused).}
}
\description{
Summary of Perfect Foresight Transition
}
